move scheduler concept into its own header
